renames package and field
