+++ src/main/java/server/routing/FunctionalHandler.java
@@ -1,9 +1,9 @@
-package server.route;
+package server.routing;
 
 import server.request.Request;
 import server.response.Response;
 
 @FunctionalInterface
-public interface RequestHandler {
+public interface FunctionalHandler {
     Response handle(Request request);
 }
