+++ src/main/java/server/routing/Routes.java
@@ -1,4 +1,5 @@
-package server.route;
+package server.routing;
 
 public class Routes {
+
 }
